# scaffold/hr-service/openapi.yaml
openapi: 3.0.3
info:
  title: PMP Human Resources Service
  version: "1.0"
  description: >
    Implements FR-HR-01..07 (docs Section 4.6). FastAPI must generate a schema
    that matches this contract — treat mismatches as a bug, not a spec update,
    unless the SRS itself is being revised.

    All endpoints require a bearer access token (RS256 JWT) issued by
    iam-service; its signature is verified against iam-service's JWKS and the
    required `<resource>.<action>` permission must be present in the token's
    `permissions` claim (CLAUDE.md rule 4). Missing/invalid token -> 401;
    valid token without the permission -> 403.

    discipline_records are additionally restricted: only hr.discipline.read /
    hr.discipline.write unlock those routes (docs Section 2.3 — HR/command
    only). FR-HR-08 (roster/headcount reports) is not implemented — it reads
    to derive reports rather than defining new domain state, and is deferred.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /officers:
    post:
      summary: Create an officer master profile (FR-HR-01)
      description: >
        Requires `hr.officer.write`. Also opens the officer's first
        assignment (FR-HR-02) at `unit_id`. Publishes an `OfficerCreated`
        domain event.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OfficerCreate" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Officer" } } } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.officer.write }
        "404": { description: unit_id does not exist }
        "409": { description: user_id or badge_number already in use }
    get:
      summary: List officers (FR-HR-01)
      description: Requires `hr.officer.read`.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: unit_id, schema: { type: string, format: uuid } }
        - { in: query, name: status, schema: { type: string, enum: [active, on_leave, suspended, retired] } }
        - { in: query, name: limit, schema: { type: integer, default: 50, minimum: 1, maximum: 200 } }
        - { in: query, name: offset, schema: { type: integer, default: 0, minimum: 0 } }
      responses:
        "200":
          description: Officers, ordered by badge number
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Officer" } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.officer.read }
  /officers/{officer_id}:
    get:
      summary: Retrieve one officer's profile (FR-HR-01)
      description: Requires `hr.officer.read`.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: officer_id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Officer" } } } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.officer.read }
        "404": { description: No officer with that id }
    patch:
      summary: Correct badge_number/hire_date/status (FR-HR-01)
      description: >
        Requires `hr.officer.write`. `rank` and `unit_id` are NOT editable
        here — they're owned by the audited promotion (FR-HR-04) and
        transfer-approval (FR-HR-03) workflows. Publishes an `OfficerUpdated`
        domain event.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: officer_id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OfficerUpdate" }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/Officer" } } } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.officer.write }
        "404": { description: No officer with that id }
        "409": { description: badge_number already in use }
...
  /officers/{officer_id}/leave-requests:
    post:
      summary: Request leave (FR-HR-05)
      description: >
        Requires `hr.leave.write`. Opens a `pending` leave request. Publishes
        a `LeaveRequested` domain event.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: officer_id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeaveRequestCreate" }
      responses:
        "201": { description: Requested, content: { application/json: { schema: { $ref: "#/components/schemas/LeaveRequest" } } } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.leave.write }
        "404": { description: No officer with that id }
    get:
      summary: Leave-request history for one officer (FR-HR-05)
      description: Requires `hr.leave.read`.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: officer_id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Leave requests, newest first
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/LeaveRequest" } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.leave.read }
        "404": { description: No officer with that id }
  /leave-requests:
    get:
      summary: Approver queue — every leave request (FR-HR-05)
      description: Requires `hr.leave.read`.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: status, schema: { type: string, enum: [pending, approved, rejected] } }
      responses:
        "200":
          description: Leave requests, newest first
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/LeaveRequest" } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.leave.read }
  /leave-requests/{leave_request_id}:
    patch:
      summary: Approve or reject a pending leave request (FR-HR-05)
      description: >
        Requires `hr.leave.approve`. Routed to the requester's supervisor per
        FR-HR-05 — enforced here as the `hr.leave.approve` permission, held by
        Station Commander and HR Officer. Publishes a `LeaveStatusChanged`
        domain event.
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: leave_request_id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeaveStatusUpdate" }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/LeaveRequest" } } } }
        "401": { description: Missing or invalid access token }
        "403": { description: Caller lacks hr.leave.approve }
        "404": { description: No leave request with that id }
        "409": { description: Leave request is not pending }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 access token issued by iam-service (/api/v1/auth/mfa/verify).
  schemas:
...
    OfficerCreate:
      type: object
      required: [user_id, badge_number, rank, unit_id, hire_date]
      properties:
        user_id: { type: string, format: uuid }
        badge_number: { type: string }
        rank: { type: string }
        unit_id: { type: string, format: uuid }
        hire_date: { type: string, format: date }
        status: { type: string, enum: [active, on_leave, suspended, retired], default: active }
    Officer:
      allOf:
        - $ref: "#/components/schemas/OfficerCreate"
        - type: object
          properties:
            id: { type: string, format: uuid }
    OfficerUpdate:
      type: object
      properties:
        badge_number: { type: string, nullable: true }
        hire_date: { type: string, format: date, nullable: true }
        status: { type: string, enum: [active, on_leave, suspended, retired], nullable: true }
...
    LeaveRequestCreate:
      type: object
      required: [leave_type]
      properties:
        leave_type: { type: string }
    LeaveRequest:
      allOf:
        - $ref: "#/components/schemas/LeaveRequestCreate"
        - type: object
          properties:
            id: { type: string, format: uuid }
            officer_id: { type: string, format: uuid }
            status: { type: string, enum: [pending, approved, rejected] }
            created_at: { type: string, format: date-time }
    LeaveStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [approved, rejected] }
...